[balance] motors can sense the state of the mpu and operate

+++ src/hal/motor.h
@@ -9,7 +9,7 @@
 typedef struct {
     bool is_outbound;
     int32_t position;
-    float angle_offset;
+    double angle_offset;
 }MotorStatusInfo;
 
 struct XKnobConfig {
